refactor fetch data, implementation repository, scheduler fetch data

+++ src/models/request/pagination_request.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 
-#[derive(Deserialize, Serialize)]
+#[derive(Deserialize, Serialize, Clone)]
 pub struct PaginationRequest {
-    pub page: Option<i32>,
-    pub limit: Option<i32>,
+    pub page: Option<u64>,
+    pub limit: Option<u64>,
 }
